// src/interfaces/propertyadvert.ts
export interface IPropertyAdvert {
    name: string;
    legal_identification: string;
    phone: string;
    email: string;
    zipcode: string;
    address: string;
    address_complement: string;
    neighborhood: string;
    city: string;
    state: string;
    country: string;
    finality: string;
    property_type: string;
    availability: string;
    sale_value: string;
    rental_value: string;
    bedroom: string;
    suite: string;
    garage: string;
    description: string;
}